feat(wasm): size_t-width ABI fix — an allocating Hale program runs in wasm, no stubs (WASM Phase 1)
The self-contained wasm runtime now not only compiles + links but RUNS:
a struct-allocating Hale program (arena alloc + field GEP) executes under
Node against the compiled-in lotus runtime with ZERO imports and ZERO JS
stubs — struct_sum() == 330. This closes the Phase-1 milestone.

The blocker was the Phase-0-flagged risk #2: codegen declared runtime
size/len args as i64 (native size_t), but wasm32 size_t is i32, so calls
traded a mismatched ABI and trapped. Fix, targeting the reachable path:

- codegen: add `usize_type()` (target-pointer-width int — i64 native /
  i32 wasm32) + `size_to_usize()` (identity on 64-bit, i64->i32 truncate
  on wasm). `memcpy` is declared with a `size_t`-width `n` and
  emit_memcpy_call narrows the count — matching the libc ABI on both
  targets (no-op on native).
- runtime: lotus_arena_alloc takes fixed-width uint64_t size/align (not
  size_t) so the FFI ABI is uniform across targets — codegen's i64 always
  matches (uint64_t == size_t on 64-bit native; stays 64-bit on wasm32).
- test: wasm_struct_runs_self_contained compiles the real runtime +
  bundled libc for wasm32, links with NO --allow-undefined / NO stubs,
  and asserts struct_sum()==330 under Node. The honest end-to-end proof
  (gated on clang/wasm-ld/node; skips cleanly without them).

Native unaffected: usize == i64 there, so the conversions are no-ops;
verified across prog (385), the ffi string/return paths, and the
bus payload-copy (memcpy) dispatch tests.

FOLLOW-UP (the remaining 4 wasm-ld signature-mismatch warnings):
lotus_bus_payload_arena_alloc / lotus_bus_dispatch / lotus_fs_* /
lotus_file_* still carry size_t params declared i64. They're unreachable
from a compute program (gc-stripped, harmless), but programs using the
bus/fs need the same uint64_t treatment — a mechanical sweep using
wasm-ld's mismatch warnings as the oracle. Plus link_wasm CLI wiring +
the _hale_start/_hale_tick entry inversion are Phase 2.

+++ crates/hale-codegen/runtime/lotus_arena.c
@@ -1713,7 +1713,11 @@ static void *lotus_arena_alloc_nolock(lotus_arena_t *a, size_t size, size_t alig
  * lotus_arena_alloc happens under the lock (the fresh-chunk path mallocs /
  * pulls from the per-thread chunk pool), so the non-recursive mutex is
  * safe. */
-void *lotus_arena_alloc(lotus_arena_t *a, size_t size, size_t align) {
+/* size/align are fixed-width uint64_t (not size_t) so the FFI ABI is
+ * uniform across targets — codegen always passes i64, which matches on
+ * 64-bit native (uint64_t == size_t) AND on wasm32 (where size_t is
+ * 32-bit but the declared param stays 64-bit). WASM plan. */
+void *lotus_arena_alloc(lotus_arena_t *a, uint64_t size, uint64_t align) {
     if (a && a->shared_concurrent) {
         pthread_mutex_lock(&a->subregion_lock);
         void *p = lotus_arena_alloc_nolock(a, size, align);
